+++ api/account_balance_report.go
@@ -12,11 +12,11 @@ func NewAccountBalanceReportService(c *Client) *AccountBalanceReportService {
 	return &AccountBalanceReportService{client: c}
 }
 
-func (svc *AccountBalanceReportService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountBalanceReportService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountBalanceReportModel, name)
 }
 
-func (svc *AccountBalanceReportService) GetByIds(ids []int) (*types.AccountBalanceReports, error) {
+func (svc *AccountBalanceReportService) GetByIds(ids []int64) (*types.AccountBalanceReports, error) {
 	a := &types.AccountBalanceReports{}
 	return a, svc.client.getByIds(types.AccountBalanceReportModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountBalanceReportService) GetAll() (*types.AccountBalanceReports, 
 	return a, svc.client.getAll(types.AccountBalanceReportModel, a)
 }
 
-func (svc *AccountBalanceReportService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountBalanceReportService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountBalanceReportModel, fields, relations)
 }
 
-func (svc *AccountBalanceReportService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountBalanceReportService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountBalanceReportModel, ids, fields, relations)
 }
 
-func (svc *AccountBalanceReportService) Delete(ids []int) error {
+func (svc *AccountBalanceReportService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountBalanceReportModel, ids)
 }
