+++ api/account_bank_statement_cashbox.go
@@ -12,11 +12,11 @@ func NewAccountBankStatementCashboxService(c *Client) *AccountBankStatementCashb
 	return &AccountBankStatementCashboxService{client: c}
 }
 
-func (svc *AccountBankStatementCashboxService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountBankStatementCashboxService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountBankStatementCashboxModel, name)
 }
 
-func (svc *AccountBankStatementCashboxService) GetByIds(ids []int) (*types.AccountBankStatementCashboxs, error) {
+func (svc *AccountBankStatementCashboxService) GetByIds(ids []int64) (*types.AccountBankStatementCashboxs, error) {
 	a := &types.AccountBankStatementCashboxs{}
 	return a, svc.client.getByIds(types.AccountBankStatementCashboxModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountBankStatementCashboxService) GetAll() (*types.AccountBankState
 	return a, svc.client.getAll(types.AccountBankStatementCashboxModel, a)
 }
 
-func (svc *AccountBankStatementCashboxService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountBankStatementCashboxService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountBankStatementCashboxModel, fields, relations)
 }
 
-func (svc *AccountBankStatementCashboxService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountBankStatementCashboxService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountBankStatementCashboxModel, ids, fields, relations)
 }
 
-func (svc *AccountBankStatementCashboxService) Delete(ids []int) error {
+func (svc *AccountBankStatementCashboxService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountBankStatementCashboxModel, ids)
 }
