+++ api/account_bank_statement_import.go
@@ -12,11 +12,11 @@ func NewAccountBankStatementImportService(c *Client) *AccountBankStatementImport
 	return &AccountBankStatementImportService{client: c}
 }
 
-func (svc *AccountBankStatementImportService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountBankStatementImportService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountBankStatementImportModel, name)
 }
 
-func (svc *AccountBankStatementImportService) GetByIds(ids []int) (*types.AccountBankStatementImports, error) {
+func (svc *AccountBankStatementImportService) GetByIds(ids []int64) (*types.AccountBankStatementImports, error) {
 	a := &types.AccountBankStatementImports{}
 	return a, svc.client.getByIds(types.AccountBankStatementImportModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountBankStatementImportService) GetAll() (*types.AccountBankStatem
 	return a, svc.client.getAll(types.AccountBankStatementImportModel, a)
 }
 
-func (svc *AccountBankStatementImportService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountBankStatementImportService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountBankStatementImportModel, fields, relations)
 }
 
-func (svc *AccountBankStatementImportService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountBankStatementImportService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountBankStatementImportModel, ids, fields, relations)
 }
 
-func (svc *AccountBankStatementImportService) Delete(ids []int) error {
+func (svc *AccountBankStatementImportService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountBankStatementImportModel, ids)
 }
