+++ api/account_invoice_line.go
@@ -12,11 +12,11 @@ func NewAccountInvoiceLineService(c *Client) *AccountInvoiceLineService {
 	return &AccountInvoiceLineService{client: c}
 }
 
-func (svc *AccountInvoiceLineService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountInvoiceLineService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountInvoiceLineModel, name)
 }
 
-func (svc *AccountInvoiceLineService) GetByIds(ids []int) (*types.AccountInvoiceLines, error) {
+func (svc *AccountInvoiceLineService) GetByIds(ids []int64) (*types.AccountInvoiceLines, error) {
 	a := &types.AccountInvoiceLines{}
 	return a, svc.client.getByIds(types.AccountInvoiceLineModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountInvoiceLineService) GetAll() (*types.AccountInvoiceLines, erro
 	return a, svc.client.getAll(types.AccountInvoiceLineModel, a)
 }
 
-func (svc *AccountInvoiceLineService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountInvoiceLineService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountInvoiceLineModel, fields, relations)
 }
 
-func (svc *AccountInvoiceLineService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountInvoiceLineService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountInvoiceLineModel, ids, fields, relations)
 }
 
-func (svc *AccountInvoiceLineService) Delete(ids []int) error {
+func (svc *AccountInvoiceLineService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountInvoiceLineModel, ids)
 }
